backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Google ADK + Gemini
google-adk>=0.3.0
google-generativeai>=0.7.2

# MongoDB — pinned together to avoid motor/beanie compatibility issues
motor==3.3.2
pymongo==4.6.3
beanie==1.25.0

# Validation & config
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# HTTP client (for external APIs)
httpx>=0.27.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
ulid-py>=1.1.0